+++ camomilla/admin.py
@@ -11,7 +11,7 @@ from hvad.admin import TranslatableAdmin
 from hvad.forms import TranslatableModelForm
 
 from django.conf import settings
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 class UserProfileAdmin(admin.ModelAdmin):