+++ camomilla/fields.py
@@ -1,11 +1,21 @@
 import json
-
+import django
 from django.conf import settings
-from django.contrib.postgres.fields import (
-    JSONField as DjangoJSONField,
-    ArrayField as DjangoArrayField,
+if django.VERSION >= (4, 0):
+    from django.db.models import JSONField as DjangoJSONField
+else:
+    from django.contrib.postgres.fields import JSONField as DjangoJSONField
+from django.contrib.postgres.fields import ArrayField as DjangoArrayField
+from django.db import models
+
+
+ORDERING_ACCEPTED_FIELDS = (
+    models.BigIntegerField,
+    models.IntegerField,
+    models.PositiveIntegerField,
+    models.PositiveSmallIntegerField,
+    models.SmallIntegerField,
 )
-from django.db.models import Field
 
 
 class JSONField(DjangoJSONField):
@@ -18,7 +28,7 @@ class ArrayField(DjangoArrayField):
 
 if "sqlite" in settings.DATABASES["default"]["ENGINE"]:
 
-    class JSONField(Field):
+    class JSONField(models.Field):
         def db_type(self, connection):
             return "text"