+++ camomilla/models/media.py
@@ -13,7 +13,7 @@ from ..fields import JSONField
 from django.db.models.signals import post_save, pre_delete
 from django.dispatch import receiver
 from django.utils.safestring import mark_safe
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from hvad.models import TranslatableModel, TranslatedFields
 from PIL import Image