+++ camomilla/serializers/fields/related.py
@@ -1,5 +1,5 @@
 from collections import OrderedDict
-from rest_framework import serializers
+from rest_framework import serializers, relations
 
 
 class RelatedField(serializers.PrimaryKeyRelatedField):
@@ -15,9 +15,11 @@ class RelatedField(serializers.PrimaryKeyRelatedField):
             ), 'Class {serializer_class} missing "Meta.model" attribute'.format(
                 serializer_class=self.serializer.__class__.__name__
             )
-            kwargs["queryset"] = kwargs.get(
-                "queryset", self.serializer.Meta.model.objects.all()
-            )
+            if not kwargs.get("read_only", False):
+                kwargs["queryset"] = kwargs.get(
+                    "queryset", self.serializer.Meta.model.objects.all()
+                )
+            self.allow_insert = kwargs.pop("allow_insert", False)
             # kwargs["allow_null"] = kwargs.get("allow_null", self.serializer.Meta.model._meta.get_field(self.source).null)
         super().__init__(**kwargs)
 
@@ -31,7 +33,18 @@ class RelatedField(serializers.PrimaryKeyRelatedField):
 
     def to_internal_value(self, data):
         if isinstance(data, dict):
-            return self.get_queryset().get(**{self.lookup: data.get(self.lookup, None)})
+            instance = (
+                self.get_queryset()
+                .filter(**{self.lookup: data.get(self.lookup, None)})
+                .first()
+            )
+            if self.allow_insert is True and len(data.keys()) and self.serializer:
+                serialized_data = self.serializer(
+                    instance=instance, data=data, context=self.context
+                )
+                if serialized_data.is_valid(raise_exception=ValueError):
+                    instance = serialized_data.save()
+            return instance
         return super().to_internal_value(data)
 
     def get_choices(self, cutoff=None):
@@ -50,3 +63,44 @@ class RelatedField(serializers.PrimaryKeyRelatedField):
                 for item in queryset
             ]
         )
+
+    @classmethod
+    def many_init(cls, *args, **kwargs):
+        class ManyRelatedField(serializers.ManyRelatedField):
+            def to_internal_value(self, data):
+                if isinstance(data, str) or not hasattr(data, "__iter__"):
+                    self.fail("not_a_list", input_type=type(data).__name__)
+                if not self.allow_empty and len(data) == 0:
+                    self.fail("empty")
+
+                child = self.child_relation
+                instances = {
+                    getattr(item, child.lookup): item
+                    for item in child.get_queryset().filter(
+                        **{
+                            f"{child.lookup}__in": [
+                                item.get(child.lookup, None)
+                                if isinstance(item, dict)
+                                else item
+                                for item in data
+                            ]
+                        }
+                    )
+                }
+                if child.allow_insert is True and len(data.keys()) and child.serializer:
+                    for item in data:
+                        serialized_data = child.serializer(
+                            instance=instances.get(item.get(child.lookup)),
+                            data=item,
+                            context=child.context,
+                        )
+                        if serialized_data.is_valid(raise_exception=ValueError):
+                            instance = serialized_data.save()
+                            instances[instance.pk] = instance
+                return instances.values()
+
+        list_kwargs = {"child_relation": cls(*args, **kwargs)}
+        for key in kwargs:
+            if key in relations.MANY_RELATION_KWARGS:
+                list_kwargs[key] = kwargs[key]
+        return ManyRelatedField(**list_kwargs)