+++ camomilla/serializers/page.py
@@ -1,6 +1,4 @@
 from ..models import Content, Page
-from .fields import RelatedField
-from .media import MediaSerializer
 from .base import BaseTranslatableModelSerializer
 
 
@@ -11,11 +9,6 @@ class ContentSerializer(BaseTranslatableModelSerializer):
 
 
 class PageSerializer(BaseTranslatableModelSerializer):
-    og_image = RelatedField(serializer=MediaSerializer, required=False, allow_null=True)
-    contents = RelatedField(
-        serializer=ContentSerializer, many=True, required=False, allow_null=True
-    )
-
     class Meta:
         model = Page
         fields = "__all__"