+++ conftest.py
@@ -1,6 +1,24 @@
 import os
 from django.conf import settings
 
+from django.core.management import call_command
+from django.db import connections
+import pytest
+
+# # UNCOMMENT THIS TO TEST WITH A LOCAL SQLITE DB FIXTURE
+
+# @pytest.fixture(scope='session')
+# def django_db_setup(django_db_blocker):
+#     from django.conf import settings
+#     test_db = 'test_db'
+#     settings.DATABASES['default']['NAME'] = test_db
+
+#     with django_db_blocker.unblock():
+#         call_command('migrate', '--noinput')
+#     yield
+#     for connection in connections.all():
+#         connection.close()
+
 
 settings.configure(
     DATABASES={"default": {"ENGINE": "django.db.backends.sqlite3"}},
@@ -25,7 +43,6 @@ settings.configure(
         ("de", "German"),
     ),
     USE_I18N=True,
-    USE_L10N=True,
     THUMB_FOLDER="thumbnails",
     REST_FRAMEWORK={
         "DEFAULT_PARSER_CLASSES": (